rename expo-plugin-pwa to expo/pwa
