+++ packages/pwa/src/fav/ico.ts
