+++ packages/pwa/src/fav/resize.ts
