+++ packages/pwa/src/generate-manifest.ts
