+++ packages/pwa/src/generateAppIconsAsync.ts
