+++ packages/pwa/src/generateFaviconsAsync.ts
