+++ packages/pwa/src/generateSplashScreensAsync.ts
