+++ packages/pwa/src/splash.ts
