+++ packages/pwa/src/update.ts
